+++ package.json
@@ -6,6 +6,7 @@
   "author": "Kyle <mkitigy@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "cypress": "6.4.0",
     "eslint": "^7.20.0",
     "eslint-config-prettier": "^7.2.0",
