+++ src/js/index.js
@@ -1,4 +1,6 @@
+import { LottoView } from './view/index.js';
 import { LottoController } from './controller/index.js';
+import { LottoMachine } from './model/index.js';
 
 class App {
   constructor() {
@@ -6,7 +8,7 @@ class App {
   }
 
   init() {
-    const controller = new LottoController();
+    const controller = new LottoController(new LottoMachine(), new LottoView());
 
     controller.initEvent();
   }
