+++ src/js/util/index.js
@@ -1,3 +1,3 @@
 export { $ } from './DOM.js';
 export { validator } from './validator.js';
-export { getRandomNumber } from './get_random_number.js';
+export { getRandomNumber } from './getRandomNumber.js';
