+++ src/js/view/LottoView.js
@@ -10,11 +10,11 @@ export class LottoView {
     $('#lotto-result-form').show();
 
     function lottoTemplate(lottos) {
-      return lottos.reduce((html, lotto) => {
+      return lottos.reduce((html, lotto, idx) => {
         return (html += ` 
           <div class="lotto-wrapper d-flex items-start">
             <span class="lotto mx-1 text-4xl">🎟️ </span>
-            <span class="lotto-numbers mx-1 text-2xl d-none">
+            <span data-lotto-numbers=${idx} class="mx-1 text-2xl d-none">
               ${lotto.numbers.join(', ')}
             </span>
           </div>
